+++ src/main/java/com/dace/dmgr/game/GameUser.java
@@ -249,7 +249,7 @@ public final class GameUser implements Disposable {
                     user.setTabListItem(column, (i + 1) * 3 - 2, UserData.fromPlayer(teamUsers[i].getPlayer()).getDisplayName(),
                             this.team == team2 || headReveal ? Skins.getPlayer(teamUsers[i].getPlayer()) : Skins.getPlayer("crashdummie99"));
                     user.setTabListItem(column, (i + 1) * 3 - 1, MessageFormat.format("§7✪ §f{0}   §7{1} §f{2}   §7{3} §f{4}   §7{5} §f{6}",
-                            teamUsers[i].getScore(), TextIcon.DAMAGE, teamUsers[i].getKill(), TextIcon.POISON, teamUsers[i].getDeath(),
+                            (int) teamUsers[i].getScore(), TextIcon.DAMAGE, teamUsers[i].getKill(), TextIcon.POISON, teamUsers[i].getDeath(),
                             "✔", teamUsers[i].getAssist()), null);
                 }
             }
